# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skillbytes API - PBL Module",
        "description": "REST API for Problem-Based Learning (PBL) Case Management System",
        "contact": {
            "email": "support@skillbytes.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development"
        },
        {
            "url": "http://145.79.13.180/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/pbl-cases": {
            "get": {
                "tags": [
                    "PBL Cases"
                ],
                "summary": "List all PBL cases",
                "description": "Retrieve all available PBL cases with current user's status",
                "operationId": "listPblCases",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "mata_pelajaran_id",
                        "in": "query",
                        "description": "Filter by mata pelajaran ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of PBL cases",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "case_number": {
                                                        "type": "integer"
                                                    },
                                                    "image_url": {
                                                        "type": "string"
                                                    },
                                                    "time_limit": {
                                                        "type": "integer"
                                                    },
                                                    "start_date": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "deadline": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "not-started",
                                                            "in-progress",
                                                            "completed",
                                                            "late"
                                                        ]
                                                    },
                                                    "mata_pelajaran": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PBL Cases"
                ],
                "summary": "Create a new PBL case (Admin/Guru only)",
                "description": "Create a new PBL case with all details. Only accessible to Admin and Guru roles.",
                "operationId": "createPblCase",
                "requestBody": {
                    "description": "PBL case data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "case_number",
                                    "title",
                                    "mata_pelajaran_id",
                                    "description",
                                    "start_date",
                                    "deadline"
                                ],
                                "properties": {
                                    "case_number": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "System Login Bermasalah"
                                    },
                                    "mata_pelajaran_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Anda diminta untuk menyelesaikan..."
                                    },
                                    "time_limit": {
                                        "type": "integer",
                                        "example": 120
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PBL case created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "slug": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Only Admin and Guru can create cases"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pbl-cases/{pblCase}": {
            "get": {
                "tags": [
                    "PBL Cases"
                ],
                "summary": "Get PBL case details",
                "description": "Retrieve full details of a specific PBL case including all sections and items",
                "operationId": "getPblCase",
                "parameters": [
                    {
                        "name": "pblCase",
                        "in": "path",
                        "description": "Case ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed PBL case information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "slug": {
                                            "type": "string"
                                        },
                                        "case_number": {
                                            "type": "integer"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "image_url": {
                                            "type": "string"
                                        },
                                        "time_limit": {
                                            "type": "integer"
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "deadline": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "not-started",
                                                "in-progress",
                                                "completed",
                                                "late"
                                            ]
                                        },
                                        "sections": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "order": {
                                                        "type": "integer"
                                                    },
                                                    "items": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "type": {
                                                                    "type": "string",
                                                                    "enum": [
                                                                        "heading",
                                                                        "text",
                                                                        "list",
                                                                        "image"
                                                                    ]
                                                                },
                                                                "content": {
                                                                    "type": "string"
                                                                },
                                                                "image_url": {
                                                                    "type": "string"
                                                                },
                                                                "order": {
                                                                    "type": "integer"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Case not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "PBL Cases"
                ],
                "summary": "Update a PBL case (Admin/Guru only)",
                "description": "Update an existing PBL case. Only accessible to Admin and Guru roles.",
                "operationId": "updatePblCase",
                "parameters": [
                    {
                        "name": "pblCase",
                        "in": "path",
                        "description": "PBL case ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "mata_pelajaran_id": {
                                        "type": "integer"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PBL case updated successfully"
                    },
                    "404": {
                        "description": "PBL case not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Only Admin and Guru can update cases"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PBL Cases"
                ],
                "summary": "Delete a PBL case (Admin/Guru only)",
                "description": "Delete an existing PBL case. Only accessible to Admin and Guru roles.",
                "operationId": "deletePblCase",
                "parameters": [
                    {
                        "name": "pblCase",
                        "in": "path",
                        "description": "PBL case ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PBL case deleted successfully"
                    },
                    "404": {
                        "description": "PBL case not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Only Admin and Guru can delete cases"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}